Environment wrappers refactoring
