+++ tests/golem/envs/test_auto_setup.py
@@ -12,7 +12,7 @@ from golem.envs import (
     RuntimeEventType,
     RuntimePayload
 )
-from golem.envs.auto_setup import auto_setup
+from golem.envs.wrappers.auto_setup import auto_setup
 
 
 class TestAutoSetup(TwistedTestCase):
@@ -30,9 +30,6 @@ class TestAutoSetup(TwistedTestCase):
         self.master_mock = Mock()  # To assert call order between multiple mocks
         self.master_mock.attach_mock(self.start_usage, 'start_usage')
         self.master_mock.attach_mock(self.end_usage, 'end_usage')
-
-    def test_supported(self):
-        self.assertEqual(self.wrapped_env.supported(), self.env.supported())
 
     def test_parse_prerequisites(self):
         prereq_dict = {'key': 'value'}
